refactor: use new ui components
