+++ nym-vpn-desktop/src/ui/index.ts
@@ -1,10 +1,16 @@
+export { default as Button } from './Button';
+export { default as MsIcon } from './MsIcon';
+export { default as RadioGroup } from './RadioGroup';
+export { default as Switch } from './Switch';
+export { default as SettingsMenuCard } from './SettingsMenuCard';
+export { default as TextArea } from './TextArea';
+export { default as TextInput } from './TextInput';
 export { default as ThemeSetter } from './ThemeSetter';
 export { default as TopBar } from './TopBar';
-export { default as Switch } from './Switch';
-export { default as RadioGroup } from './RadioGroup';
-export { default as MsIcon } from './MsIcon';
-export { default as SettingsMenuCard } from './SettingsMenuCard';
-export * from './Switch';
-export * from './RadioGroup';
+export * from './Button';
 export * from './MsIcon';
+export * from './RadioGroup';
 export * from './SettingsMenuCard';
+export * from './Switch';
+export * from './TextArea';
+export * from './TextInput';
